FLP-1493 script now acceptance test environment friendly

+++ src/Database/Scripts/PreDeploymentScript.sql
@@ -32,20 +32,49 @@ GO
 --FLP-1493 DELETE AFTER
 BEGIN TRAN
 
-DELETE FROM [Domain].[MathsAndEnglishPeriodInLearning];
-DELETE FROM [Domain].[MathsAndEnglishInstalment];
-DELETE FROM [Domain].[MathsAndEnglish];
+IF EXISTS (SELECT 1 FROM sys.tables t JOIN sys.schemas s ON t.schema_id = s.schema_id 
+           WHERE t.name = 'MathsAndEnglishPeriodInLearning' AND s.name = 'Domain')
+    DELETE FROM [Domain].[MathsAndEnglishPeriodInLearning];
 
-DELETE FROM [Domain].[Instalment];
-DELETE FROM [Domain].[AdditionalPayment];
-DELETE FROM [History].[EarningsProfileHistory];
-DELETE FROM [Domain].[EarningsProfile];
+IF EXISTS (SELECT 1 FROM sys.tables t JOIN sys.schemas s ON t.schema_id = s.schema_id 
+           WHERE t.name = 'MathsAndEnglishInstalment' AND s.name = 'Domain')
+    DELETE FROM [Domain].[MathsAndEnglishInstalment];
 
-DELETE FROM [Domain].[EpisodePrice];
-DELETE FROM [Domain].[EpisodePeriodInLearning];
-DELETE FROM [Domain].[Episode];
+IF EXISTS (SELECT 1 FROM sys.tables t JOIN sys.schemas s ON t.schema_id = s.schema_id 
+           WHERE t.name = 'MathsAndEnglish' AND s.name = 'Domain')
+    DELETE FROM [Domain].[MathsAndEnglish];
 
-DELETE FROM [Domain].[Apprenticeship];
+IF EXISTS (SELECT 1 FROM sys.tables t JOIN sys.schemas s ON t.schema_id = s.schema_id 
+           WHERE t.name = 'Instalment' AND s.name = 'Domain')
+    DELETE FROM [Domain].[Instalment];
+
+IF EXISTS (SELECT 1 FROM sys.tables t JOIN sys.schemas s ON t.schema_id = s.schema_id 
+           WHERE t.name = 'AdditionalPayment' AND s.name = 'Domain')
+    DELETE FROM [Domain].[AdditionalPayment];
+
+IF EXISTS (SELECT 1 FROM sys.tables t JOIN sys.schemas s ON t.schema_id = s.schema_id 
+           WHERE t.name = 'EarningsProfileHistory' AND s.name = 'History')
+    DELETE FROM [History].[EarningsProfileHistory];
+
+IF EXISTS (SELECT 1 FROM sys.tables t JOIN sys.schemas s ON t.schema_id = s.schema_id 
+           WHERE t.name = 'EarningsProfile' AND s.name = 'Domain')
+    DELETE FROM [Domain].[EarningsProfile];
+
+IF EXISTS (SELECT 1 FROM sys.tables t JOIN sys.schemas s ON t.schema_id = s.schema_id 
+           WHERE t.name = 'EpisodePrice' AND s.name = 'Domain')
+    DELETE FROM [Domain].[EpisodePrice];
+
+IF EXISTS (SELECT 1 FROM sys.tables t JOIN sys.schemas s ON t.schema_id = s.schema_id 
+           WHERE t.name = 'EpisodePeriodInLearning' AND s.name = 'Domain')
+    DELETE FROM [Domain].[EpisodePeriodInLearning];
+
+IF EXISTS (SELECT 1 FROM sys.tables t JOIN sys.schemas s ON t.schema_id = s.schema_id 
+           WHERE t.name = 'Episode' AND s.name = 'Domain')
+    DELETE FROM [Domain].[Episode];
+
+IF EXISTS (SELECT 1 FROM sys.tables t JOIN sys.schemas s ON t.schema_id = s.schema_id 
+           WHERE t.name = 'Apprenticeship' AND s.name = 'Domain')
+    DELETE FROM [Domain].[Apprenticeship];
 
 COMMIT TRAN
 GO
